track call stack depth of multiplot commands
Replace the flag
    TBOOLEAN multiplot
with
    int in_multiplot

Motivation:

   The commands making up a multiplot are saved to $GPVAL_LAST_MULTIPLOT
so that the plots can be regenerated for mousing or replotting.
The new variable in_multiplot tracks what the stack depth was at the time
"set multiplot" is executed.  This will allow us to suppress saving any
command executed at a deeper stack depth so that "load 'foo'" is saved but
the content of file 'foo' is not.