improved multiplot mousing (5/4) - better approach to saving multiplot commands
The reverted commit 5970433a7 fixed many failure cases for remultiplot,
but not all. Furthermore it introduced new failure cases.

Try again with a different approach to saving the commands that
are executed while creating a multiplot.

Simple commands (anything without a bracketed clause) are first
executed, then saved to $GPVAL_LAST_MULTIPLOT.

if/else/do/while commands are instead saved prior to execution, including the
bracketed clauses. In this case saving to $GPVAL_LAST_MULTIPLOT is suppressed
during the execution so that you don't end up with, say, ten copies of the
same loop.

load/call commands are treated similarly to if/do/while.  The load command
itself is saved but not the individual commands inside the loaded file.