qt:  Work-around for a strange bug in toggling plots on/off
Reproducer:
	set term qt
	set key opaque
	plot sin(x), atan(x), cos(atan(x)) title "BUG"

Plots 1 and 2 toggle correctly, either interactively by clicking on
the key text or by typing "toggle 1" or "toggle 2" as expected.
Plot 3 does not respond to the key text, and can only be toggled
from the command line by "toggle 4".
I have not been able to figure out why the numbering is bad only
for qt and only for the final plot in a sequence.

The opaque key triggers a second pass over the plots, so that is
no doubt part of it.  But  wxt/x11/svg do not show this problem.

Work-around:
	Open and close an empty plot in draw_key() before the second
	pass if the terminal is qt.  Not satisfactory but it works.

Bug 2761