Enable the HIDDEN3d_CACHE option by default
HIDDEN3D_CACHE is an experimental option added in 2023 (commit 438a66cc1d)
that tries to reconcile the drawing order of quadrangles rendered by the
pm3d "depthorder" code and lines drawn in splot's "hidden3d" mode.
It works but there didn't seem to be any great motivation or need to make
it the default.

However it turns out that something changed on the way to version 6
that broke attempts to draw contours onto a depth-ordered pm3d surface
(Bug 2762).  Since the HIDDEN3D_CACHE code handles this case, I think
it is time to enable it by default and consider it for the next
6.0 release.

Bug 2762