save axis mappings across a "reset" command
After a reset command the plot that is showing on the screen is no
longer mousable because reset has clobbered the axis mapping values
(notably axis->min and axis->max but also axis->term_lower axis->term_upper).
If we save these beforehand just as we do for the active plot region
then mousing can use them after the reset.  This leaves the previous plot
(the one still showing on the screen in an interactive session) still
able to report mouse coordinates. However keystroke events (e.g. 'bind'
and replot) are affected by the reset and any subsequent commands.

The saved parameters are not sufficient to handle nonlinear or linked axes
in general, but simple log scaling is handled.

This change demonstrates that in priniciple saving this set of values for
each plot in a multiplot would allow subsequent mousing of all the 2D panels.
However it is not obvious where to store per-panel information for a multiplot.