revert assignment of LT_NODRAW to variable color with value = NaN
This reverts commit a9e47d5a4
Setting the linetype to LT_NODRAW has the desired effect of not drawing
a line segment with color = NaN, but has the unfortunate side effect that
the calling sites do not realize that the linetype has changed.
So the previous linetype does not get restored and subsequent line segments
are skipped also.

The reverts to setting the color to LT_BACKGROUND, but also adds a
check for color=NAN in plot_lines.

Bug 2772