tabulate variable color column for splot
In most cases "set table; splot ..." dumps only x/y/z to the output.
If the plot actually requires four columns of input data the fourth
column should be dumped also.

This commit does not cover all possible cases, but it catches the
most common one where the fourth column is used for variable color.

The extra column is printed using the tic format of the color axis.

Example:
   set table "out"
   set cbtics format "%.3f"
   splot DATA using 1:2:3:4 with points lc palette
   unset table