hsteps should take defaults from "set boxwidth"
This commit modifies the hsteps style so that when the width is not
provided by the third column of input data, it defaults to the width
set by 'set boxwidth'. If boxwidth is not set, the conventional method
of determining the width will be applied.

The motivation for this change is that, when certain filters
(such as bins or smooth frequency) are used in combination with the hsteps,
the third column specifying the width may not be available.
In the case of the boxes, the same limitation exists, but the width can be
specified using 'set boxwidth' command.
I wanted to enable the same behavior for the 'hsteps'.

This commit adds support for both boxwidth relative and boxwidth absolute.

This change does not affect the steps, fsteps, histeps, or fillsteps styles, even though they share the "plot_hsteps()" implementation.