"set boxwidth" is now applied to "plot ... bins ... with hsteps"
Commit 22feb809b2 introduced changes allowing the "hsteps" style to use
the "set boxwidth" setting, but it did not fully address cases where the
`bins` filter is applied.

This commit modifies the behavior so that even when the `bins` filter is used,
the "hsteps" style can still respect the width setting specified by "set boxwidth".

When the `bins` filter is applied, the number of members in each bin is stored
in "points[i].z". However, "plot_hsteps()" expects "points[i].z" to represent
the width of each segment. Since the `bins` filter provides invalid data in
"points[i].z" as the width, this caused incorrect handling of "set boxwidth".

This commit ensures that when the `bins` filter is applied, "plot_hsteps()"
no longer uses "points[i].z" for determining the segment width.