OK to have missing corners in a image represented by a sparse matrix
When a sparse matrix is populated, missing entries are flagged UNDEFINED.
However we do know their indices, and this is sufficient to generate
coordinates for plotting.  The image rendering code does a quick check
to see if the first pixel of an image is UNDEFINED and fails if so.
We don't want this to be the case for a sparse matrix with a missing corner,
so make sure the corners are flagged INRANGE even if the content is NaN
because the pixel is missing.

Bug 2791