win: FARPROC loaded by name from a dll requires a template function prototype
gcc15 fails to compile winmain with error
    winmain.c:287:14: error: too many arguments to function 'pSHGetSpecialFolderPath';
    expected 0, have 4

As I understand it, recent Windows versions have repurposed
the old FARPROC type to be a generic function pointer type that is used
as for all routines found by name by querying a dll.

However FARPROC implies a function prototype    UINT func(void),
i.e. a function with no arguments.
gcc15 then treats it as an error when the code later invokes this function
with its proper number of parameters, whatever that might be.
Earlier gcc versions at worst reported this as a warning, not an error.
Perhaps this is required by the C23 standard?

To avoid this error the expected true function prototype can be
defined as a typedef and used as a cast.  See
https://learn.microsoft.com/en-us/cpp/build/getprocaddress?view=msvc-170

Bug 2794