provide a wrapper remultiplot_command() for replay_multiplot()
When a multiplot is replayed as a result of a "replot" or "remultiplot"
command, c_token (the command line parsing pointer) must increment.
This is now done by remultiplot_command() rather than by the internal
routine replay_multiplot().

When the multiplot is replayed internally, e.g. from a mousing
operation or execution of a string, c_token should not be altered.

Bug 2796