prevent use-after-free caused by redefining an active function
It is legal to define a function inside a function block.
But if the function itself then calls that function block either directly
or indirectly, the previous definition is freed even though it is in the
process of being executed.  This causes a use-after-free error.

To prevent this we introduce a depth counter in the structure that holds
the function definition (at_type).  When the function is evaluated this
counter is incremented beforehand and decremented afterwards.

An attempt to redefine a function whose depth counter is currently greater
than 0 is rejected as an error.

Bug 2797