Catch reassignment of a global array inside a function block
Suppose there is a global array A[].  Any assignment statement
A = foo will free the data in A.  If this happens inside a function block
there is a possibility that it corrupts an operation on A in the caller.
For example, suppose we have
        function $func() << EOF
                A = 1
        EOF
        array A[3]
        A[1] = $func(x)

If the assignment A=1 is allowed to destroy and free array A, then pointers
into the original array structure that are still on the execution stack
become invalid and memory corruption or an immediate segfault results.

Bug 2798