report error on reassignment of a global array inside a function
Previous commits in this series addressed corruption of an array
structure as a side-effect of evaluating a function blocks during
an assignment operation.  This commit addresses the equivalent
side-effects of evaluating a user function.

However, there is a caveat regarding this fix. When using the action table
for evaluation, it is not possible to obtain information about the
left-hand side (i.e., the destination array) before evaluating the
right-hand side. As a result, we can accurately detect changes such as
an array no longer being an array, but not when one array is replaced
with another.
Even so, in the case of array-to-array replacement, fortunately,
the execution can proceed without triggering a fatal error.

Bug 2798