Revise the ARRAY structure (1/2) - Add a new field array_header {.size; .parent}
NB: This is the fourth attempt to address Bug 1798
    It merges the first 2 commits of version 3 with the udv->refcount commits
    from 6.0 branch

This is the first commit in a series the revises the ARRAY bookkeeping structures
and uses the revised information to catch various corruption errors.
It is an alternaive to commits
    77204ab57 Catch reassignment of a global array inside a function block
    4e3d9cc80 Invalid "stack underflow" error in variable assignment expression
    aa95c3063 origin/HEAD) Detect side-effect of array assignment that would corrupt the array
    904b362a4 report error on reassignment of a global array inside a function
These would have to be reverted prior to merging this alternative

The physical layout of the udv for an ARRAY is the same, but there is now a
new aliased field array_header in "struct value"

udvt_entry for array A
    .udv_name  "A"
    .udv_value
	.type ARRAY
	.v
	    .value_array [0:size]
		.type ARRAY | TEMP_ARRAY | COLORMAP_ARRAY | LOCAL_ARRAY
		.v
		    .array_header
			.size
			.parent (points back to udvt_entry)

The size of array "A" was previously accessed as
	udv->udv_value.v.value_array[0].v.int_val
Now the same slot in the structure is accessed as
	udv->udv_value.v.value_array[0].v.array_header.size

Aliasing to array_header gains an extra field, parent, that can be used to point
back to the udvt_entry.  This could be used to retrieve the array name as
	value_array[0].v.array_header.parent->udv_name
parent == NULL indicates that the array is not a named user variable,
e.g. it has been dynamically created during evaluation.

Importantly, it can also be used to implement reference counting for array access
to address the class of bugs described in bugs 2798 and 2802 in which a possible
side-effect of an array assignment statement A=foo or A[i]=foo is to invoke a
function or function block that overwrites the udv associated with "A".

Bug 2798