array interlock (1/5) new interlock field udv->udv_refcount in struct udvt_entry
Currently used only to prevent memory corruption resulting from an
ARRAY user variable being overwritten or redefined as something else
while an array operation is in progress.

Each array operation increments array->udv_refcount at the start
and decrements in when finished.  If the reference count is greater than 0
then assignment to the array udv is not permitted.

For example, the operation A[i] = f(x) increments the refcount for A
at the start so that while f(x) is being evaluated certain operations
are prohibited.  Examples of statements that would be rejected:
    A = 42
    array A[N]      # however local array A[N] is OK
    array B=A; A=B

The reference count must be maintained and checked in two contexts.
1) array operations given at the command line via "load"

2) array operations executed during evaluation of a function

Bug 2798