array interlock (4/5) use LOCK / UNLOCK in parse_assignment_expression
This is the easy case, where we know at the time of building
an action table which array it is that we want to protect.
The LOCK/UNLOCK operations take a variable name, although at this
time we only really care about arrays.
The sequence of operations is now:

gnuplot> show at A = f(x)
        lock A
        pushc "A"
        push x
        call f (undefined)
        assign
        unlock A

Array A is protected against being clobbered by any side effects of
evaluating f(x).