array interlock (5/5) use array interlock in places that do not call LOCK / UNLOCK
Assignment operations from the command line or in a function block
do not proceed by first building an action table, so the LOCK/UNLOCK
operations are not called.  Instead we increment/decrement the
array reference count directly.

It turns out the this leaves only a single call site for UNLOCK.
All other uses have been folded into the ASSIGN operation, which means that
the count returns to zero as expected just before the variable is updated.