Use interlock to protect an array passed as a parameter to a user function
When an array is passed as a function parameter, it is passed by reference.
If a global variable holding the array referenced by the function parameter
is overwritten during the function evaluation, it results in the array
pointer being freed, leading to an “accessing freed memory” error.
This is a variation of Bug #2798.

The following script demonstrates this issue.
The array stored in variable "A" is passed to the function parameter "x",
and "A" is overwritten during the function's execution:

function $print(var) <<end
  print var
end

f(x) = (A=1, $print(x), x)

array A = ["my", "original", "array"]
print f(A)

Bug 2802