array interlock (cleanup)
The array interlock does not protect the array during caculation of
an array index. E.g.
    A[ f(x) ] = y
fails badly if f(x) somehow clobbers A.  This sequence of events can
(probably) only happen inside a function or function block.

There is not a easy fix for this, but at least we can prevent a segfault.