colorbox: fix rendering of inverted transparent colorbox
An inverted colorbox is drawn using a sequence of filled rectangles.
The colorbox_draw_polygon() routine failed to take the inversion
into account when setting the rectangle coordinates.
This resulted in overlapping boxes.
This was not noticeable for solid fill, but overlapping transparent
boxes give muddled colors.

    set colorbox invert
    set style fill transparent solid 0.25

This fix does not affect non-inverted colorboxes, which some
terminals treat as an image rather than a set of boxes.
Possible problems with a trasparent non-inverted colorbox will
require a separate fix.