colorbox: do not render colorbox as image if fillstyle is not FS_OPAQUE
Some terminals (notably qt) can render the colorbox as an image rather
than as a set of filled rectangles.  However the image rendering code
pays no attention to fill style, so if the fill is anything other than
FS_OPAQUE this introduces a terminal-dependent difference in the output.

This commit suppresses the difference by always using the filled
rectangle method if the default fillstyle is something other than
FS_OPAQUE.

Arguably a better fix would be to give the colorbox its own fillstyle
and use that rather than the default.