Fix regression in use of named palettes for "splot with pm3d"
The regression, introduced by commit fc98188b6, only manifests when a
named palette is used for splot "with pm3d fc palette" and no line properties
precede the fillcolor in the splot command. In this circumstance the
TC_COLORMAP type and colormap are correctly read but stored as a
line property rather than a fill color.

This commit reverts deletion of a test in lp_parse() that would prevent
this error.

named_palettes.dem is modified so that the regression and fix are
clearly visible when run as a unit test.

Bug 2809