colorspec (1/7) Add field colorspec.rgbcolor via anonymous union
Rationale:

The existing field colorspec.lt is used to hold both signed integers
(e.g. LT_BLACK = -1) and unsigned integers (32bit ARGB values).
This causes type-conversion problems when the top bit of an alpha value
is set.   E.g. half-transparent blue might be represented as
0x7F0000FF (alpha = 127) or 0x800000FF (alpha = 128).
In the former case (top bit not set) everything works.

However if the top bit is set, x86 and ARM platforms behave differently.
If the 32-bit value was originally held in a signed integer variable,
it looked like a negative number and thus was stored as a negative
floating point value in the (double).  If this is later assigned to
an unsigned int variable, ARM may truncate the value to zero.

Conversely if the 32-bit value was originally held as an unsigned
integer and thus stored as a positive value, later assignment to
a signed integer variable overflows and ARM may truncate the value
to MAXINT.

Solution:

typedef struct t_colorspec {
    colortype type;             /* TC_<type> definitions below */
    union {
        int lt;                 /* used for TC_LT, TC_LINESTYLE */
        unsigned rgbcolor;      /* used for TC_RGB */
    };
    double value;               /* used for TC_CB and TC_FRAC */
} t_colorspec;

Now we can do instead colorspec.rgbcolor = point[i].CRD_COLOR
(double -> unsigned int) which preserves all 32 bits AARRGGBB.

Note:

This commit changes only the t_colorspec structure definition and
defined constants.  Changes to the code will come later.

There are two advantages to adding rgbcolor as part of an anonymous
union rather than as a separate new field.  It saves space overall
(not a huge concern but I'm an old fossil that hates wasting memory).
It means that any places in the code that were missed will continue
to act as they always have, whereas putting lt and rgbcolor in separate
places would guarantee breakage wherever the wrong one was used.
In fact I used that to debug the code changes initially, but I might
have missed some cases or unusual combinations of input options.