colorspec (2/7) use t_colorspec.rgbcolor for fillcolor in 3D boxes and polygons
Now that there is a colorspec field that explicitly holds a 32bit ARGB value
as an unsigned integer, use it to pass fillcolor for boxes and polygons.

There are many places in the code that should probably use the new field
rather than t_colorspec.lt.  This commit addresses only two cases where
the signed/unsigned difference was shown to cause a problem on ARM.

Bug 2812