nibbles: replace eval statements with arrays and function blocks
The game was written at a time when gnuplot did not offer support
for arrays, function blocks, or local variables. It hence had to
revert to ingenious use of strings and 'eval' statements. With
gnuplot 6, we can now let this look more sane.

Note that the array coll[] now stores the snake's location.