expand Unicode escape sequences \U+xxxx in "noenhanced" strings
The unicode escape sequence processing added in version 5.4 was only
applied in "enhanced text" mode.  Now the processing is applied earlier
so that it is in effect whether on not enhanced text mode is active.

It is safest to put a space after the escape code so that a 4-character
code point is not mistaken as a 5-character code point when the following
character is a legal hexdigit.
In enhanced text mode the escape sequence can be wrapped in curly brackets
to avoid this ambiguity
	e.g. write α1+α2 as {\U+03B1}1+{\U+03B1}2
For "noenhance" text you must add a space
	e.g. write α1+α2 as \U+03B1 1+\U+03B1 2