Expand unicode escape sequnces during substring evaluation
Unicode escape codes are stored in a string without interpretation,
and later replaced by the character they represent when the string
is printed or evaluated during substring processing.

Thus after defining a string S = ’A + \U+03A3 B’, S[5:5] evaluates to "Σ"
rather than "\", but (S eq S[1:*]) evaluates to FALSE because S holds
the unprocessed escape sequence while evaluation of S[1:*] replaces the
escape sequence with the UTF-8 byte sequence for Σ.
Both will display as "A + Σ B" when used in a plot.

NB: This only applies when the encoding is utf8