docs:  Modify for better use with lualatex
Recent texlive pdflatex has problems with some UTF-8 characters in
the documentation.  At some point I think we will have to make lualatex
the default, but for now just tweak the formatting somewhat to make
the User Manual look a bit better when processed with lualatex.

The Release Notes for 6.0 already suggest
    PDFLATEX=lualatex ./configure