docs: reorganize and correct errors in description of format specifiers
I messed up when revising the section on format specifiers so that
it included description of new gprint specifiers (commit 4f4f190a).
This commit attempts to disentangle documentation of gprint as a
function from the numerical format specifiers that it uses.
It tries to distinguish more clearly the three categories of
specifier:  numeric, time/date, geographic