revisit escape sequences (UTF-8)
UTF-8 escape sequences may contain either 4 or 5 hex characters
\U+hhhh or \U+hhhhh.  This can cause problems if the escape sequence
is immediately following by another hex character.
Suppose the desired output is "α1 + α2"
In enhanced text mode this can be dealt with by adding brackets:
    A = "{\U+03B1}1 + {\U+03B1}2"
For "noenhance" text this isn't an option.
Before the best work-around was to add extra spaces.
The obvious alternative
    A = "\U+03B1\1 + \U+03B1\2"
did not work because the single-character escape was applied first,
leaving an apparently valid 5-character unicode sequence.

Now the nonenhanced text alternative works because we catch this case
on input in util.c: parse_esci() and replace the '\' with the ascii unit
separator '\037'.  When the unicode sequence is later expanded
term.c: enhanced recursion() and expand_unicode_escapes()
we strip this separator out again.