kitty: support animation on a transparent background
If the background is set to be transparent then usually we want to
clear the previous contents before making a new plot.
Multiplots are an exception, since the whole idea is to add new content
without erasing the old content.

This allows animation on a transparent background without superposition
of successive frames.

    set term kitty transparent
    load 'animate2.dem'