webp: support animation on a transparent background
Because the webp terminal piggybacks on cairo+png to create images,
the png image must be re-initialized at the start of each frame.
Otherwise successive images superimpose on each other rather than
sequentially replacing each other.

NB:
   The resulting transparent background for animations seems to work
when viewed in vwebp but is flaky when viewed in the ImageMagick
"animate" utility.