pm: reset flag term_force_init for each plot
The flag term_force_init was introduced originally for the pm terminal,
which set it once and then assumed it would stay that way.
The kitty and webp terminals now also use this flag to force
reinitialization between frames of an animation.
However for this use it is cleared after use at the start of a plot
and then reset prior to the next plot.

This commit adds the reset for pm also, so that the flag is not left
as FALSE when the pm terminal wants it to be TRUE.