Transparent sixelgd option doesn't work
This is because the Sixel background select parameter was not correctly set.
The corrected setting is
	fprintf(out_fp, "\033P9;1q\"1;1;%d;%d\n", map_width, map_height);

The first parameter (the 9) sets the aspect ratio to 1:1.
This isn't strictly necessary, because it's also later set by the raster
attributes command, but this ensures a fallback if that command isn't supported.

The second parameter (the 1) tells the terminal not to fill the background,
so when you omit pixels from an image, they'll actually be transparent.
This is the essential change that should fix the bug.

I've confirmed that this patch gets transparency working on several Sixel
terminals, including Foot, Contour, MLTerm, WezTerm, Windows Terminal, and XTerm.

It also fixes an issue on some terminals where you get an unwanted black bar
at the bottom of the image (this is an undesirable side effect of the
background fill that can occur when an image is scrolled).