unconditionally include <stdbool.h>
gcc15 is now defaulting to -std=c23
This caused build configuration/build errors in release 6.0.3 because
c23 implements
	bool true false
as reserved keywords, which breaks the autoconf test AC_HEADER_STDBOOL
for versions of autoconf prior to autoconf 2.72.

On the other hand we have raised the minimal compiler support to c99
and c99 guarantees to provide stdbool.h.
So rather than introducing additional checks and fallbacks for
autoconf versioning and Boolean type support, it should be safe
to unconditionally include <stdbool.h>.

Bug #2830