Make pre-defined variables pi NaN I and Inf read-only
Introduce a convension that udv->locality == -1 indicates the variable
is a pre-defined constant that cannot be undefined or over-written.

An attempt to overwrite the value will create a new local variable with
the same name.  This will shadow the pre-defined variable until it is
removed or until the next "reset session".
The local variable can be deleted with "undefine <varname>".

E.g.
    pi = 3.0
    show var pi
       Variables beginning with pi:
        pi = 3.0
        pi = 3.14159265358979
    undefine pi
    show var pi
        Variables beginning with pi:
        pi = 3.14159265358979

Notes:

  The rationale for allowing the user to shadow a pre-defined variable
  is that I is a plausible name for a normal user variable in contexts
  where complex arithmetic is not used.

  However it would be possible to silently ignore an attempt to
  overwrite pi or Inf while still allowing use of I as a local variable.
  I.e.
    locality -1	allow shadowing
    locality -2	disallow shadowing and ignore attempts to change the value